Implement pricing strategies and availability observer

+++ src/main/java/org/example/vehicle/Car.java
@@ -6,12 +6,12 @@ public class Car implements Vehicle{
     private final String id;
     private final String model;
     private final String engine;
-    private final int seats;
-    private final int basePricePerDay;
-    private final int basePricePerHour;
+    private final float seats;
+    private final float basePricePerDay;
+    private final float basePricePerHour;
 
-    public Car(String id, String model, String engine, int seats,
-               int basePricePerDay, int basePricePerHour) {
+    public Car(String id, String model, String engine, float seats,
+               float basePricePerDay, float basePricePerHour) {
         this.id = Objects.requireNonNull(id);
         this.model = Objects.requireNonNull(model);
         this.engine = engine;
@@ -23,8 +23,8 @@ public class Car implements Vehicle{
     @Override public String getId() { return id; }
     @Override public String getType() { return "Car"; }
     @Override public String getModel() { return model; }
-    @Override public int getBasePricePerDay() { return basePricePerDay; }
-    @Override public int getBasePricePerHour() { return basePricePerHour; }
+    @Override public float getBasePricePerDay() { return basePricePerDay; }
+    @Override public float getBasePricePerHour() { return basePricePerHour; }
     @Override public String getDescription() {
         return String.format("Car[id=%s, model=%s, engine=%s, seats=%d]", id, model, engine, seats);
     }
